lint: satisfy ruff 0.16 and pin it so CI can't drift from local
CI installs the newest ruff (was >=0.6, unpinned) and 0.16.0 introduced
rules the local 0.15 run didn't flag. Apply its fixes across src and
tests (import sorting, collections.abc Callable, context-managed lock
file, narrowed except in the table-enable probe, f-string for the
wait_for_response description, subprocess.run check=False in tests)
and pin ruff==0.16.0 in the dev extra so the gate is deterministic.

+++ src/inkbridge/convert/targeted.py
@@ -43,10 +43,10 @@ def _bbox_to_pixels(
     """
     h, w = shape
     nx, ny, nw, nh = bbox_norm
-    x0 = int(round(nx * w)) - pad_px
-    y0 = int(round(ny * h)) - pad_px
-    x1 = int(round((nx + nw) * w)) + pad_px
-    y1 = int(round((ny + nh) * h)) + pad_px
+    x0 = round(nx * w) - pad_px
+    y0 = round(ny * h) - pad_px
+    x1 = round((nx + nw) * w) + pad_px
+    y1 = round((ny + nh) * h) + pad_px
     x0 = max(0, min(x0, w))
     x1 = max(0, min(x1, w))
     y0 = max(0, min(y0, h))
